+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## Unreleased
+
+### Features
+
+- Add support for building on Windows ([#7](https://github.com/getsentry/sentry-maven-plugin/pull/7))
+
+### Fixes
+
+- Also use `url` and `authToken` for upload command ([#7](https://github.com/getsentry/sentry-maven-plugin/pull/7))
+
 ## 0.0.2
 
 ### Features